# Scalable-friendly dct-inl.h