@eustas 2025-01-08
(no changes, just now tested to be performance neutral)